refactor(food-search): 共有検索フォーム作成
