+++ src/app/editor-meal/food-search/food-search.component.ts
@@ -1,18 +1,10 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { FoodService } from 'src/app/services/food.service';
 import { AuthService } from 'src/app/services/auth.service';
-import { DailyInfoService } from 'src/app/services/daily-info.service';
+
 import { ActivatedRoute, Router, NavigationStart } from '@angular/router';
-import { SearchService } from 'src/app/services/search.service';
-import { Food } from 'src/app/interfaces/food';
-import { DailyMeal } from 'src/app/interfaces/daily-info';
+
 import { Subscription } from 'rxjs';
 import { AverageService } from 'src/app/services/average.service';
-import { Validators, FormBuilder, FormArray } from '@angular/forms';
-import { QueryDocumentSnapshot } from '@angular/fire/firestore';
-import { take } from 'rxjs/operators';
-import { MatBottomSheet } from '@angular/material/bottom-sheet';
-import { MealInputComponent } from 'src/app/bottom-sheet/meal-input/meal-input.component';
 import { MainShellService } from 'src/app/services/main-shell.service';
 
 @Component({
@@ -22,46 +14,19 @@ import { MainShellService } from 'src/app/services/main-shell.service';
 })
 export class FoodSearchComponent implements OnInit, OnDestroy {
   private subscription = new Subscription();
-  private date: string;
-
+  date: string;
   readonly maxSelectNum = 50;
-  readonly getNumber = 500;
-  readonly minAmount = 0;
-  readonly maxAmount = 10000;
   selectedMealsNum: number;
-  amount = new Array(10).fill(0);
-  isLikedlist: string[] = new Array();
 
-  readonly index = this.searchService.index;
-  amountForm = this.fb.group({
-    amountArray: this.fb.array([]),
-  });
-  get amountArray(): FormArray {
-    return this.amountForm.get('amountArray') as FormArray;
-  }
-  loading: boolean;
-
-  hitFoods: Food[] = [];
-
-  readonly form = this.fb.group({
-    query: [''],
-  });
   constructor(
-    private foodService: FoodService,
     private authService: AuthService,
-    private dailyInfoService: DailyInfoService,
     private route: ActivatedRoute,
-    private searchService: SearchService,
     private router: Router,
     private averageService: AverageService,
-    private fb: FormBuilder,
-    private bottomSheet: MatBottomSheet,
     private mainShellService: MainShellService
   ) {
     const routeSub = this.route.queryParamMap.subscribe((paramMaps) => {
       this.date = paramMaps.get('date');
-      this.getFavFoods();
-      this.setFormArray();
     });
 
     const selectedSub = this.getSelectedMeals();
@@ -75,72 +40,6 @@ export class FoodSearchComponent implements OnInit, OnDestroy {
     this.subscription.add(selectedSub);
     this.subscription.add(routerSub);
   }
-  private setFormArray() {
-    for (let i = 0; i < 10; i++) {
-      const amountGroup = this.fb.group({
-        [i]: [
-          0,
-          [Validators.min(this.minAmount), Validators.max(this.maxAmount)],
-        ],
-      });
-      this.amountArray.push(amountGroup);
-    }
-  }
-
-  private getFavFoods() {
-    this.foodService
-      .getFavFoods(this.authService.uid, this.getNumber)
-      .pipe(take(1))
-      .subscribe((foods) => {
-        if (foods && foods.length > 0) {
-          foods.forEach(
-            (food: {
-              data: Food;
-              nextLastDoc: QueryDocumentSnapshot<Food>;
-            }) => {
-              this.isLikedlist.push(food.data.foodId);
-            }
-          );
-        }
-      });
-  }
-
-  likeFavFood(food: Food) {
-    this.foodService.likeFavFood(this.authService.uid, food);
-    this.isLikedlist.push(food.foodId);
-  }
-
-  unLikeFavFood(foodId: string) {
-    this.foodService.unLikeFavFood(this.authService.uid, foodId);
-    const index = this.isLikedlist.indexOf(foodId);
-    if (index > -1) {
-      this.isLikedlist.splice(index, 1);
-    }
-  }
-
-  addFood(amount: number, food: Food) {
-    if (amount >= 0) {
-      const meal: DailyMeal = { mealId: '', food, amount };
-      this.dailyInfoService.addMeal(
-        meal,
-        this.authService.uid,
-        this.date,
-        'food'
-      );
-    }
-  }
-
-  openBottomSheet(food: Food): void {
-    this.bottomSheet.open(MealInputComponent, {
-      data: {
-        userId: this.authService.uid,
-        date: this.date,
-        maxAmount: this.maxAmount,
-        minAmount: this.minAmount,
-        food,
-      },
-    });
-  }
 
   private getSelectedMeals() {
     this.mainShellService.selectedMeals.subscribe((v) => {
@@ -148,29 +47,7 @@ export class FoodSearchComponent implements OnInit, OnDestroy {
     });
   }
 
-  ngOnInit(): void {
-    this.search('こめ');
-  }
-
-  submitQuery(): void {
-    this.search(this.form.value.query);
-  }
-
-  private search(query: string): void {
-    this.loading = true;
-    this.index
-      .search(query, {
-        hitsPerPage: 10,
-      })
-      .then(({ hits }) => {
-        this.hitFoods = (hits as unknown) as Food[];
-        this.loading = false;
-      });
-  }
-
-  clearForm() {
-    this.form.reset();
-  }
+  ngOnInit(): void {}
 
   ngOnDestroy() {
     this.subscription.unsubscribe();
