+++ src/app/food-search/food-list-for-meal/food-list-for-meal.component.html
@@ -1,0 +1,88 @@
+<p class="limit-fav" *ngIf="isLikedlist.length === getNumber">
+  ＊これ以上お気に入りに登録できません
+</p>
+<div *ngFor="let food of foods; index as i">
+  <div class="food">
+    <button
+      class="food__icon"
+      mat-icon-button
+      color="warn"
+      *ngIf="isLikedlist.includes(food.foodId)"
+      (click)="unLikeFavFood(food.foodId)"
+    >
+      <mat-icon>favorite</mat-icon>
+    </button>
+
+    <button
+      class="food__icon"
+      mat-icon-button
+      *ngIf="!isLikedlist.includes(food.foodId)"
+      (click)="likeFavFood(food)"
+      [disabled]="isLikedlist.length === getNumber"
+    >
+      <mat-icon>favorite_border</mat-icon>
+    </button>
+    <div class="food__name">
+      <p class="food__name--with-amount">{{ food.foodName }}</p>
+      <span class="food__amountper">
+        {{ food.foodCalPerAmount }}kcal/{{ food.unit }}
+      </span>
+    </div>
+    <div class="spacer"></div>
+
+    <form class="food__choice" [formGroup]="amountForm">
+      <div formArrayName="amountArray">
+        <mat-form-field
+          class="food__choice-form-field"
+          appearance="outline"
+          [formGroupName]="i"
+        >
+          <input
+            matInput
+            class="food__choice-input"
+            autocomplete="off"
+            [formControlName]="i"
+            [max]="maxAmount"
+            [min]="minAmount"
+            step="0.1"
+            [(ngModel)]="amount[i]"
+            type="number"
+          />
+          <mat-hint align="end">
+            {{ food.foodCalPerAmount * amount[i] | number: '1.1-1' }}
+            kcal</mat-hint
+          >
+          <span matSuffix>{{ food.unit }}</span>
+          <mat-error *ngIf="amount[i] > maxAmount || amount[i] < minAmount"
+            >{{ minAmount }}〜{{ maxAmount }}の範囲で入力してください</mat-error
+          >
+        </mat-form-field>
+      </div>
+    </form>
+
+    <button
+      class="open-sheet"
+      mat-icon-button
+      color="primary"
+      (click)="openBottomSheet(food)"
+      [disabled]="isAddable"
+    >
+      <mat-icon>add_box</mat-icon>
+    </button>
+
+    <button
+      class="add-meal"
+      mat-icon-button
+      color="primary"
+      (click)="addFood(amount[i], food)"
+      [disabled]="
+        amount[i] === null ||
+        amount[i] > maxAmount ||
+        amount[i] < minAmount ||
+        isAddable
+      "
+    >
+      <mat-icon>add_box</mat-icon>
+    </button>
+  </div>
+</div>
