+++ src/app/food-search/food-list-for-meal/food-list-for-meal.component.scss
@@ -1,0 +1,66 @@
+.food {
+  display: flex;
+  align-items: center;
+  padding: 16px 0;
+
+  &__fav {
+    margin: 0 8px 0 0px;
+  }
+
+  &__amountper {
+    font-size: 12px;
+    opacity: 0.7;
+    margin: 0;
+    color: #009688;
+  }
+
+  &__name {
+    font-size: 14px;
+    margin-right: 8px;
+    width: 40%;
+
+    @media screen and(max-width:750px) {
+      width: 100%;
+    }
+
+    &--with-amount {
+      margin: 0;
+    }
+  }
+
+  &__choice {
+    display: flex;
+    flex-direction: column;
+    margin-right: 8px;
+    width: 30%;
+
+    @media screen and(max-width:750px) {
+      display: none;
+    }
+  }
+
+  &__choice-form-field {
+    font-size: 14px;
+    margin-bottom: 4px;
+  }
+
+  &__choice-input {
+    display: block;
+    width: 95%;
+    font-size: 16px;
+    text-align: right;
+  }
+
+  &__cal {
+    margin: 0;
+    text-align: right;
+    font-size: 14px;
+    opacity: 0.87;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+}
+
+.limit-fav {
+  color: #e25241;
+}
