+++ src/app/food-search/food-search-content/food-search-content.component.html
@@ -1,0 +1,19 @@
+<app-food-search-form (setFoods)="_setFoods($event)"></app-food-search-form>
+
+<ng-container [ngSwitch]="pageTitle">
+  <app-food-list-for-meal
+    *ngSwitchCase="'mealPage'"
+    [foods]="foods"
+    [date]="date"
+    [isAddable]="isAddable"
+  ></app-food-list-for-meal>
+
+  <app-food-list-for-set *ngSwitchCase="'setPage'"></app-food-list-for-set>
+</ng-container>
+
+<div *ngIf="foods?.length === 0 && !loading">
+  該当する食べ物はありません
+</div>
+<div class="loading" *ngIf="loading">
+  <mat-spinner class="loding__spinner" diameter="30"></mat-spinner>
+</div>
