+++ src/app/food-search/food-search-content/food-search-content.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { FoodSearchContentComponent } from './food-search-content.component';
+
+describe('FoodSearchContentComponent', () => {
+  let component: FoodSearchContentComponent;
+  let fixture: ComponentFixture<FoodSearchContentComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [FoodSearchContentComponent],
+    }).compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(FoodSearchContentComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
