+++ src/app/food-search/food-search-form/food-search-form.component.scss
@@ -1,0 +1,34 @@
+.form {
+  position: relative;
+  margin-bottom: 24px;
+
+  &__search {
+    position: absolute;
+    left: 16px;
+    top: 10px;
+    font-size: 20px;
+  }
+
+  &__close {
+    position: absolute;
+    right: 8px;
+    top: 2px;
+    font-size: 16px;
+    height: 32px;
+    width: 32px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+}
+.search {
+  width: 88%;
+  border-radius: 8px;
+  border: 1px solid #ddd;
+  background: #fff;
+  padding: 8px 40px 8px 32px;
+  display: block;
+  margin: 0 auto;
+  color: rgba(0, 0, 0, 0.87);
+  appearance: none;
+}
